Push markdown files to git remote after sync
ChronDB's native push was sending its internal bare repo (metadata) to
the remote instead of the actual markdown files users expect.

Moved git operations to the sync output directory so `roam sync` commits
and pushes the markdown pages/daily notes directly. ChronDB stays local
for state tracking only, git CLI handles the remote via SSH.
