+++ Cargo.toml
@@ -50,7 +50,7 @@ tree-sitter-yaml = "0.7"
 tree-sitter-md = "0.3"
 rmcp = { version = "1.1.0", features = ["server", "transport-io", "schemars"] }
 clap = { version = "4", features = ["derive"] }
-chrondb = { version = "0.1.1", optional = true }
+chrondb = { version = "0.2.3-dev.24", optional = true }
 
 [features]
 default = ["sync"]
