+++ src/sync/store.rs
@@ -1,5 +1,5 @@
 use std::collections::HashSet;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 use serde_json::{json, Value};
 
@@ -15,24 +15,74 @@ pub struct PageEntry {
 }
 
 pub struct SyncStore {
-    db_dir: PathBuf,
+    db: chrondb::ChronDB,
     /// Cached page UIDs loaded once at startup.
     known_pages: HashSet<String>,
-    /// Last sync timestamp (epoch millis), loaded from ChronDB.
+    /// Last sync timestamp (epoch millis).
     last_sync_ms: i64,
-    /// Timestamp to save on next flush (set before sync starts).
+    /// Timestamp to save on next flush.
     next_sync_ms: i64,
-    /// Pending writes — flushed to ChronDB in one batch at the end.
+    /// Pending writes — flushed in one batch at the end.
     pending: Vec<PageEntry>,
 }
 
 impl SyncStore {
     pub fn open(db_dir: &Path) -> Result<Self> {
-        let db_dir = db_dir.to_path_buf();
-        let (known_pages, last_sync_ms) = Self::load_state(&db_dir);
+        let data_path = db_dir.join("data");
+        let index_path = db_dir.join("index");
+
+        std::fs::create_dir_all(&data_path)
+            .map_err(|e| RoamError::Generic(format!("Failed to create data dir: {}", e)))?;
+        std::fs::create_dir_all(&index_path)
+            .map_err(|e| RoamError::Generic(format!("Failed to create index dir: {}", e)))?;
+
+        #[allow(deprecated)]
+        let db = match chrondb::ChronDB::open(
+            data_path.to_str().unwrap_or(""),
+            index_path.to_str().unwrap_or(""),
+        ) {
+            Ok(db) => db,
+            Err(_) => {
+                eprintln!("ChronDB: index corrupted, rebuilding...");
+                let _ = std::fs::remove_dir_all(&index_path);
+                let _ = std::fs::create_dir_all(&index_path);
+                #[allow(deprecated)]
+                chrondb::ChronDB::open(
+                    data_path.to_str().unwrap_or(""),
+                    index_path.to_str().unwrap_or(""),
+                )
+                .map_err(|e| RoamError::Generic(format!("Failed to open ChronDB: {}", e)))?
+            }
+        };
+
+        // Load known pages from index
+        let mut known_pages = HashSet::new();
+        if let Ok(list) = db.list_by_prefix("page:", None) {
+            if let Some(arr) = list.as_array() {
+                for item in arr {
+                    if let Some(key) = item.get("id").and_then(|v| v.as_str()) {
+                        if let Some(uid) = key.strip_prefix("page:") {
+                            known_pages.insert(uid.to_string());
+                        }
+                    }
+                }
+            } else if let Some(obj) = list.as_object() {
+                for key in obj.keys() {
+                    if let Some(uid) = key.strip_prefix("page:") {
+                        known_pages.insert(uid.to_string());
+                    }
+                }
+            }
+        }
+
+        let last_sync_ms = db
+            .get("meta:last_sync_ms", None)
+            .ok()
+            .and_then(|v| v.get("ts")?.as_i64())
+            .unwrap_or(0);
 
         Ok(Self {
-            db_dir,
+            db,
             known_pages,
             last_sync_ms,
             next_sync_ms: 0,
@@ -40,94 +90,10 @@ impl SyncStore {
         })
     }
 
-    fn data_path(&self) -> PathBuf {
-        self.db_dir.join("data")
-    }
-
-    fn index_path(&self) -> PathBuf {
-        self.db_dir.join("index")
-    }
-
-    fn load_state(db_dir: &Path) -> (HashSet<String>, i64) {
-        let data_path = db_dir.join("data");
-        let index_path = db_dir.join("index");
-
-        if !data_path.exists() {
-            return (HashSet::new(), 0);
-        }
-
-        let db = match chrondb::ChronDB::open(
-            data_path.to_str().unwrap_or(""),
-            index_path.to_str().unwrap_or(""),
-        ) {
-            Ok(db) => db,
-            Err(_) => {
-                // Index might be corrupted (e.g. from Ctrl+C). Reset and retry.
-                eprintln!("ChronDB: index corrupted, rebuilding...");
-                let _ = std::fs::remove_dir_all(&index_path);
-                let _ = std::fs::create_dir_all(&index_path);
-                match chrondb::ChronDB::open(
-                    data_path.to_str().unwrap_or(""),
-                    index_path.to_str().unwrap_or(""),
-                ) {
-                    Ok(db) => db,
-                    Err(_) => return (HashSet::new(), 0),
-                }
-            }
-        };
-
-        let mut pages = HashSet::new();
-        if let Ok(list) = db.list_by_prefix("page:", None) {
-            if let Some(arr) = list.as_array() {
-                for item in arr {
-                    if let Some(key) = item.get("id").and_then(|v| v.as_str()) {
-                        if let Some(uid) = key.strip_prefix("page:") {
-                            pages.insert(uid.to_string());
-                        }
-                    }
-                }
-            } else if let Some(obj) = list.as_object() {
-                for key in obj.keys() {
-                    if let Some(uid) = key.strip_prefix("page:") {
-                        pages.insert(uid.to_string());
-                    }
-                }
-            }
-        }
-
-        // Load last sync timestamp
-        let last_sync_ms = db
-            .get("meta:last_sync_ms", None)
-            .ok()
-            .and_then(|v| v.get("ts")?.as_i64())
-            .unwrap_or(0);
-
-        (pages, last_sync_ms)
-    }
-
-    fn open_db(&self) -> Result<chrondb::ChronDB> {
-        let data_path = self.data_path();
-        let index_path = self.index_path();
-
-        std::fs::create_dir_all(&data_path)
-            .map_err(|e| RoamError::Generic(format!("Failed to create ChronDB data dir: {}", e)))?;
-        std::fs::create_dir_all(&index_path).map_err(|e| {
-            RoamError::Generic(format!("Failed to create ChronDB index dir: {}", e))
-        })?;
-
-        chrondb::ChronDB::open(
-            data_path.to_str().unwrap_or(""),
-            index_path.to_str().unwrap_or(""),
-        )
-        .map_err(|e| RoamError::Generic(format!("Failed to open ChronDB: {}", e)))
-    }
-
-    /// Fast in-memory check — no ChronDB round-trip.
     pub fn has_page(&self, uid: &str) -> bool {
         self.known_pages.contains(uid)
     }
 
-    /// Queue a page for batch write. Updates in-memory cache immediately.
     pub fn put_page(&mut self, uid: &str, title: &str, file_path: &str, block_count: usize) {
         self.known_pages.insert(uid.to_string());
         self.pending.push(PageEntry {
@@ -138,10 +104,8 @@ impl SyncStore {
         });
     }
 
-    /// Flush pending pages to ChronDB + always save sync timestamp.
+    /// Flush pending pages + save timestamp + push to remote. Reuses the open connection.
     pub fn flush(&mut self) -> Result<()> {
-        let db = self.open_db()?;
-
         if !self.pending.is_empty() {
             eprintln!("Writing {} pages to ChronDB...", self.pending.len());
             for entry in &self.pending {
@@ -152,20 +116,20 @@ impl SyncStore {
                     "file_path": entry.file_path,
                     "block_count": entry.block_count,
                 });
-                if let Err(e) = db.put(&key, &doc, None) {
+                if let Err(e) = self.db.put(&key, &doc, None) {
                     eprintln!("ChronDB: failed to write page {}: {}", entry.uid, e);
                     break;
                 }
             }
         }
 
-        // Save sync timestamp (must be JSON object, not scalar)
+        // Save sync timestamp
         if self.next_sync_ms > 0 {
-            let _ = db.put("meta:last_sync_ms", &json!({"ts": self.next_sync_ms}), None);
+            let _ = self
+                .db
+                .put("meta:last_sync_ms", &json!({"ts": self.next_sync_ms}), None);
             self.last_sync_ms = self.next_sync_ms;
         }
-
-        drop(db);
 
         let count = self.pending.len();
         self.pending.clear();
@@ -176,12 +140,11 @@ impl SyncStore {
         Ok(())
     }
 
-    /// Get version history for a page.
     #[allow(dead_code)]
     pub fn page_history(&self, uid: &str) -> Result<Value> {
-        let db = self.open_db()?;
         let key = format!("page:{}", uid);
-        db.history(&key, None)
+        self.db
+            .history(&key, None)
             .map_err(|e| RoamError::Generic(format!("ChronDB history error: {}", e)))
     }
 
@@ -193,10 +156,7 @@ impl SyncStore {
         self.last_sync_ms
     }
 
-    /// Set the timestamp to save on next flush (call before sync starts).
     pub fn set_next_sync_ms(&mut self, ms: i64) {
         self.next_sync_ms = ms;
     }
-
-    // --- Git remote operations on ChronDB's data dir ---
 }
